Implement `TryFrom<Value>` for `f64` ☕️

+++ fennec-cli/src/api/modbus/sma.rs
@@ -6,9 +6,7 @@ pub struct Client(pub modbus::Client);
 
 impl Client {
     pub async fn read_total_export(&self) -> Result<KilowattHours> {
-        #[expect(clippy::cast_precision_loss)]
-        let watt_hours = u64::try_from(self.0.read_value().await?)? as f64;
-
+        let watt_hours = f64::try_from(self.0.read_value().await?)?;
         Ok((watt_hours / 1000.0).into())
     }
 }
